Add tests for the selection view
