+++ apps/explorer/tests/views/test_views.py
@@ -18,12 +18,10 @@ from apps.core.management.commands.make_development_fixtures import (
 )
 from apps.explorer.views import (
     PixelSetDetailView, PixelSetExportView, PixelSetExportPixelsView,
-    DataTableDetailView, DataTableSelectionView,
+    DataTableDetailView,
 )
 from apps.explorer.views.helpers import get_selected_pixel_sets_from_session
-from apps.explorer.views.views_detail import (
-    GetOmicsUnitsMixin as GetOmicsUnitsDetailMixin
-)
+from apps.explorer.views.views_detail import GetOmicsUnitsMixin
 
 
 class PixelSetListViewTestCase(CoreFixturesTestCase):
@@ -1209,7 +1207,7 @@ class PixelSetExportViewTestCase(CoreFixturesTestCase):
         )
 
 
-class PixelSetDetailViewTestCase(GetOmicsUnitsDetailMixin,
+class PixelSetDetailViewTestCase(GetOmicsUnitsMixin,
                                  CoreFixturesTestCase):
 
     def setUp(self):
@@ -1529,7 +1527,7 @@ class PixelSetExportPixelsViewTestCase(CoreFixturesTestCase):
             )
 
 
-class PixelSetDetailValuesViewTestCase(GetOmicsUnitsDetailMixin,
+class PixelSetDetailValuesViewTestCase(GetOmicsUnitsMixin,
                                        CoreFixturesTestCase):
 
     def setUp(self):
@@ -1616,7 +1614,7 @@ class PixelSetDetailValuesViewTestCase(GetOmicsUnitsDetailMixin,
         self.assertEqual(rows[0]['c'][1]['v'], selected_pixel.value)
 
 
-class PixelSetDetailQualityScoresViewTestCase(GetOmicsUnitsDetailMixin,
+class PixelSetDetailQualityScoresViewTestCase(GetOmicsUnitsMixin,
                                               CoreFixturesTestCase):
 
     def setUp(self):
@@ -1713,234 +1711,3 @@ class DataTableDetailViewTestCase(TestCase):
         with pytest.raises(NotImplementedError):
             view = DataTableDetailViewWithNoGetHeaders()
             view.get_headers()
-
-
-class PixelSetSelectionViewTestCase(CoreFixturesTestCase):
-
-    def setUp(self):
-
-        self.user = factories.PixelerFactory(
-            is_active=True,
-            is_staff=True,
-            is_superuser=True,
-        )
-        self.client.login(
-            username=self.user.username,
-            password=factories.PIXELER_PASSWORD,
-        )
-        self.url = reverse('explorer:pixelset_selection')
-
-    def test_redirects_to_list_view_when_invalid(self):
-
-        response = self.client.get(self.url)
-
-        self.assertRedirects(response, reverse('explorer:pixelset_list'))
-
-    def test_displays_message_after_redirect_when_selection_is_empty(self):
-
-        response = self.client.get(self.url, follow=True)
-
-        self.assertContains(
-            response,
-            (
-                '<div class="message error">'
-                'Cannot explore an empty selection.'
-                '</div>'
-            ),
-            html=True
-        )
-
-    def test_renders_pixelset_selection_template(self):
-
-        # select 2 pixel sets
-        pixel_sets = factories.PixelSetFactory.create_batch(2)
-        data = {
-            'pixel_sets': [str(p.id) for p in pixel_sets]
-        }
-        self.client.post(
-            reverse('explorer:pixelset_select'), data, follow=True
-        )
-
-        response = self.client.get(self.url)
-
-        self.assertEqual(response.status_code, 200)
-        self.assertTemplateUsed(response, 'explorer/pixelset_selection.html')
-
-        self.assertContains(
-            response,
-            '<title>Pixel Sets - Your selection</title>'
-        )
-
-        self.assertContains(
-            response,
-            '<div class="pixelset-item">',
-            count=len(pixel_sets)
-        )
-
-
-class DataTableSelectionViewTestCase(TestCase):
-
-    def test_get_headers_must_be_implemented(self):
-
-        class DataTableSelectionViewWithNoGetHeaders(DataTableSelectionView):
-            pass
-
-        with pytest.raises(NotImplementedError):
-            view = DataTableSelectionViewWithNoGetHeaders()
-            view.get_headers()
-
-
-class PixelSetSelectionCumulativeQualityScoresViewTestCase(CoreFixturesTestCase):  # noqa
-
-    def setUp(self):
-
-        self.user = factories.PixelerFactory(
-            is_active=True,
-            is_staff=True,
-            is_superuser=True,
-        )
-        self.client.login(
-            username=self.user.username,
-            password=factories.PIXELER_PASSWORD,
-        )
-
-        self.pixel_set = factories.PixelSetFactory()
-        self.pixels = factories.PixelFactory.create_batch(
-            2,
-            pixel_set=self.pixel_set
-        )
-
-        self.url = reverse(
-            'explorer:pixelset_selection_cumulative_quality_scores'
-        )
-
-    def test_returns_bad_request_when_not_ajax(self):
-
-        response = self.client.get(self.url)
-
-        self.assertEqual(response.status_code, 400)
-
-    def test_returns_json(self):
-
-        # select 1 pixel set
-        data = {
-            'pixel_sets': [self.pixel_set.id]
-        }
-        self.client.post(
-            reverse('explorer:pixelset_select'), data, follow=True
-        )
-
-        response = self.client.get(
-            self.url,
-            data={},
-            HTTP_X_REQUESTED_WITH='XMLHttpRequest'
-        )
-
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response['Content-Type'], 'application/json')
-
-        data = json.loads(response.content)
-
-        cols = data['cols']
-        self.assertEqual(cols[0]['label'], 'id')
-        self.assertEqual(cols[1]['label'], 'quality_score')
-
-        rows = data['rows']
-        self.assertEqual(len(rows), 2)
-
-    def test_no_selected_pixel_sets_returns_empty(self):
-
-        response = self.client.get(
-            self.url,
-            data={},
-            HTTP_X_REQUESTED_WITH='XMLHttpRequest'
-        )
-
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response['Content-Type'], 'application/json')
-
-        data = json.loads(response.content)
-
-        cols = data['cols']
-        self.assertEqual(cols[0]['label'], 'id')
-        self.assertEqual(cols[1]['label'], 'quality_score')
-
-        rows = data['rows']
-        self.assertEqual(len(rows), 0)
-
-
-class PixelSetSelectionCumulativeValuesViewTestCase(CoreFixturesTestCase):
-
-    def setUp(self):
-
-        self.user = factories.PixelerFactory(
-            is_active=True,
-            is_staff=True,
-            is_superuser=True,
-        )
-        self.client.login(
-            username=self.user.username,
-            password=factories.PIXELER_PASSWORD,
-        )
-
-        self.pixel_set = factories.PixelSetFactory()
-        self.pixels = factories.PixelFactory.create_batch(
-            2,
-            pixel_set=self.pixel_set
-        )
-
-        self.url = reverse('explorer:pixelset_selection_cumulative_values')
-
-    def test_returns_bad_request_when_not_ajax(self):
-
-        response = self.client.get(self.url)
-
-        self.assertEqual(response.status_code, 400)
-
-    def test_returns_json(self):
-
-        # select 1 pixel set
-        data = {
-            'pixel_sets': [self.pixel_set.id]
-        }
-        self.client.post(
-            reverse('explorer:pixelset_select'), data, follow=True
-        )
-
-        response = self.client.get(
-            self.url,
-            data={},
-            HTTP_X_REQUESTED_WITH='XMLHttpRequest'
-        )
-
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response['Content-Type'], 'application/json')
-
-        data = json.loads(response.content)
-
-        cols = data['cols']
-        self.assertEqual(cols[0]['label'], 'id')
-        self.assertEqual(cols[1]['label'], 'value')
-
-        rows = data['rows']
-        self.assertEqual(len(rows), 2)
-
-    def test_no_selected_pixel_sets_returns_empty(self):
-
-        response = self.client.get(
-            self.url,
-            data={},
-            HTTP_X_REQUESTED_WITH='XMLHttpRequest'
-        )
-
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response['Content-Type'], 'application/json')
-
-        data = json.loads(response.content)
-
-        cols = data['cols']
-        self.assertEqual(cols[0]['label'], 'id')
-        self.assertEqual(cols[1]['label'], 'value')
-
-        rows = data['rows']
-        self.assertEqual(len(rows), 0)
